requirements.txt
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
paho-mqtt>=1.6.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
brotli>=1.0.0
toutsurmoneau>=0.0.27
aiohttp>=3.13.0
